+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.23 (pre-release):
+* [Change] `KSP 1.8` compatibility. __WARNING__: the mod won't work with version lower than `KSP 1.8`!
 * [Fix #339] Inflatable habitat parts have unreasonable inventories.
 * [Fix #341] AddPodInventories not called for initial crew capacity 0.